+++ src/GraphQLQueryBuilder.ts
@@ -1,6 +1,6 @@
 import { GraphQLResolveInfo } from "graphql";
 import {
-  FieldNodeInfo,
+  EjectQueryCallback,
   GraphQLEntityFields,
   QueryPagination,
   QueryPredicates,
@@ -22,12 +22,14 @@ export class GraphQLQueryBuilder<T extends typeof BaseEntity> {
   private _pagination?: QueryPagination;
   private _parser: GraphQLInfoParser = new GraphQLInfoParser();
   private _context: any;
+  private _ejectQueryCallback: EjectQueryCallback<T> | null = null;
 
   constructor(
     private _manager: GraphQLQueryManager,
     private _entity: Function | string,
     private _alias?: string
-  ) {}
+  ) {
+  }
 
   /**
    * Provide the query builder with the GraphQL Query info you would like to resolve
@@ -118,13 +120,14 @@ export class GraphQLQueryBuilder<T extends typeof BaseEntity> {
    * @returns GraphQLQueryBuilder
    */
   public orWhere(
-    where: string,
+    where: WhereArgument,
     params?: ObjectLiteral
   ): GraphQLQueryBuilder<T> {
-    this._orWhereExpressions.push({
-      condition: where,
-      params
-    });
+    if (typeof where === "string") {
+      this._orWhereExpressions.push({ condition: where, params });
+    } else {
+      this._orWhereExpressions.push(where);
+    }
     return this;
   }
 
@@ -277,6 +280,44 @@ export class GraphQLQueryBuilder<T extends typeof BaseEntity> {
   }
 
   /**
+   * Receives a callback that can be used to modify the TypeORM SelectQueryBuilder instance
+   * before the loader executes the database query.
+   *
+   * Please note that this callback is run AFTER the loader has already applied all provided conditions to the query builder
+   * (where conditions, pagination, order, etc). Be aware, as changes you make to the ejected query builder could conflict
+   * with loader applied settings. Some tips to avoid potential conflicts:
+   *
+   * - If you are using the eject callback to apply where conditions, move all of your where conditions to the callback.
+   *   This will prevent potential conflicts between where conditions applied via the loader wrapper being overwritten
+   *   by conditions applied via the eject callback. Keep all your where conditions in one place.
+   *
+   * - For most cases, if you plan on joining tables in the callback, be sure to [join WITHOUT selecting](https://typeorm.io/#/select-query-builder/join-without-selection).
+   *   This is to prevent issues with selecting data from the same table twice, and is generally more performant.
+   *   If you are wanting a relation or column to always be joined and selected, see the {@link ConfigureLoader} Decorator
+   *
+   * @example
+   * ```typescript
+   * function resolve(obj, args, context, info) {
+   *   return context
+   *     .loader
+   *     .loadEntity(User, "user")
+   *     .info(info)
+   *     .ejectQueryBuilder(qb => {
+   *       return qb.innerJoin("user.group", "group")
+   *          .where("group.name = :groupName", { groupName: args.groupName })
+   *     })
+   *     .loadMany()
+   * }
+   * ```
+   *
+   * @param cb
+   */
+  public ejectQueryBuilder(cb: EjectQueryCallback<T>): GraphQLQueryBuilder<T> {
+    this._ejectQueryCallback = cb;
+    return this;
+  }
+
+  /**
    * Load one record from the database.
    * This record will include all relations and fields requested
    * in the GraphQL query that exist on the TypeORM entities.
@@ -349,13 +390,11 @@ export class GraphQLQueryBuilder<T extends typeof BaseEntity> {
   private async _genericLoad<U extends boolean, V extends boolean>(
     many: U,
     paginate: V
-  ): Promise<
-    V extends true
-      ? [InstanceType<T>[], number]
-      : U extends true
+  ): Promise<V extends true
+    ? [InstanceType<T>[], number]
+    : U extends true
       ? InstanceType<T>[]
-      : InstanceType<T> | undefined
-  > {
+      : InstanceType<T> | undefined> {
     // we need to validate an info object
     this._validateInfo(this._info);
     // Check if this query is already in the cache
@@ -384,7 +423,8 @@ export class GraphQLQueryBuilder<T extends typeof BaseEntity> {
         entity: this._entity,
         pagination: paginate ? this._pagination : undefined,
         alias: this._alias,
-        context: this._context
+        context: this._context,
+        ejectQueryCallback: this._ejectQueryCallback ?? (qb => qb)
       });
     };